Remove unused references and don't link DafnyRuntime into DafnyDriver
This is in preparation for linking the output of the Dafny-in-Dafny compiler
into Dafny (as a compiler plugin). We can't do that today because
Dafny.exe (compiled from DafnyDriver) already links in a copy of the runtime,
even though it doesn't use it.

In the long run we probably want to stop including DafnyRuntime.cs textually in
Dafny-generated programs (and instead link against a versioned DLL, at least as
an option), but that will be a breaking change for our customers and it will
require quite a bit of thinking about proper library-distribution strategies,
whereas this is a cheap and easy workaround that also happens to be the "right
thing".