fix: Substitute for receiver in automatically generated induction hypothesis (#2002)
The induction hypothesis generated for a lemma gives the lemma's postcondition for all "smaller" values of the parameters. To set this up, the verifier generates a quantifier over those "smaller" values. Previously, the receiver parameter in this quantifier was that of the lemma itself, not the bound variable in the quantifier. This PR fixes that.

Fixes #2001

Co-authored-by: Fabio Madge <fabio@madge.me>