fix: Make verifier understand (!new) (#1935)
Make the verifier understand that a type marked with `(!new)` is considered allocated in any state.

Also, this PR fixes a bug in the compilation of iterators to C#, JavaScript, and Go. The bug had caused malformed code for in-parameters of an iterator, in the case when the type of such an in-parameter was not an auto-init type.