feat: /runAllTests option for executing all {:test} methods  (#2005)
Implemented by generating most of a `Main` method in the Dafny AST, with a small amount of target language code for recovering from halting (e.g. a failed `expect` statement). Recovering from halting is not part of the Dafny language, but already implemented internally for most target languages in order to support printing `[Program Halted] ...` at the top level.

Not yet implemented for C++ or Python as they don't yet support `expect` statements.