Broaden the Ref Man authors
Though the first draft of the reference manual was written by Richard Ford and the subsequent major revision of it was written by David Cok, the reference manual has since received contributions for many others. Moreover, in additional to a team-wide effort to fill in TODOs in the manual, we are now updating the reference manual when source change require such updates. Rather than trying to list every contributor, it seems appropriate to treat the reference manual as having the same authors as the general contributors to the source code.