chore: Factor out reusable workflow for integration tests (#2077)
Progress on #2010.

This has no effect on the CI for PRs yet (aside from cleaning up the matrix configuration a bit), but introduces a "nightly build" that runs the integration tests on all platforms as well. This will be used in future changes to have the PR CI only run these tests for one platform, and having the nightly build to run the more expensive matrix instead.

Merging this change first has the benefit of allowing the nightly build to run a few times first, so that the subsequent change can add the automation to check for a successful nightly build before allowing a PR to be merged or a release to be published.