fix: reads function of seq/multiset (#2183)
This PR fixes two issues with advanced reads clauses.

One issue was with `reads F`, where `F` is an expression of type `... ~> seq<object>` or `... ~> multiset<object>`. Previously, these generated malformed Boogie. This is now fixed.

The other issue was with `reads F` where `F` is an expression of type `... ~> object`. Previously, the resolver allowed this, but the translator generated malformed Boogie. There's no strong use case for allowing such a type, so this PR fixes the issue by making the resolver forbid this case.