Feat: Plugin's Compiler/Verification errors should not prevent verification from running. (#2166)
This PR adds the ability for the Dafny Language Server to continue to verification when only verification and compilation errors are raised in the resolution phase. Only plugins would have the ability to raise verification or compilation errors in the resolution phase anyway, so this makes the plugins able to decide whether they create blocking errors or non-blocking errors.