Fixed a CI instability on hover (#2256)
Will prevent build failures such as this one:
https://github.com/dafny-lang/dafny/runs/6887347120?check_suite_focus=true

The reason was, we did hover test on the variable x of assert P(x). The problem is, with the newest changes, hovering x will also hover the assertion, so the hovering message can be more complicated.

This PR removes the ambiguity by replacing assertions with variable assignments.