docs: Add matrix of supported features by target language (#2268)
Adds a section to the User's Guide under Compilation, documenting known cases of specific backends not supporting given Dafny features. See dry run on my fork here to verify that the links and footnotes work correctly: https://robin-aws.github.io/dafny/DafnyRef/DafnyRef#2498-supported-features-by-target-language

Partially addresses #1770 - I generated the data for this table by implementing the mechanism described there (or at least a simpler version of it) on this branch, but that implementation is not yet ready, especially since it involves changing the way many if not all integration tests are run. In the meantime, having this information documented is very useful!

Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>