Optimize shift/rotate for more efficient solving
When translating `x << y` to Boogie, where `x` has width `w`, instead of
immediately casting `y` to the type of `x`, instead cast `y` to a
bitvector with width `log2(w)` and then cast, a second time, to a
bitvector of width `w`. This ensures that the solver knows that most of
the bits of the shift amount are zero.

Fixes #2174
Fixes #2230