fix: Correctly substitute type variables in inheritance checks
When Dafny translates a class, the override check that it performs mentions the
old ensures clauses, applied to its new arguments.  For example, the following
code:

```dafny
trait T           { method M(a: int) returns (b: int) ensures b > a     }
class C extends T { method M(a: int) returns (b: int) ensures b > a + 1 }
```

… leads to the following check:

```boogie
implementation OverrideCheck(this: ref, a#0: int) returns (b#0: int) {
  assume b#0 > a#0 + 1;
  assert b#0 > a#0;
}
```

Problems pop up when argument types differ between the old and new ensures
clauses, which happens in the presence of type parameters:

```dafny
predicate P<A>(s: seq<A>)
trait T<A> {
  method M(a: A) ensures P([a][0 := a])
}

predicate Q<A>(s: seq<A>)
class C extends T<object> {
  method M(a: object) ensures Q([a][0 := a])
}
```

When translating the original ensures clause (`ensures P([a][0 := a])`) to Boogie
in the override check, Dafny uses the old types in some places, and hence
assumes that `[a]` has type `seq<A>`, not type `seq<object>`.  This causes it to
make a mistake about whether to insert a box, and generate the following Boogie
code:

```boogie
assume _module.__default.Q(Tclass._System.object(),
  Seq#Update(Seq#Build(Seq#Empty(): Seq Box, $Box(a#0)), LitInt(0), $Box(a#0)));
assert _module.__default.P(_module.T$A,
  Seq#Update(Seq#Build(Seq#Empty(): Seq Box, $Box(a#0)), LitInt(0), a#0));
```

Notice how the `assert` line differs from the `assume` one: the former (with the
`Q`) correctly boxes `a#0` twice: once in the sequence construction and once in
the update.  In contrast, the latter (with `P`, corresponding to the trait's
ensures), incorrectly omits the box on the second `a#0` (the argument to
`Seq#Update`).

The reason for the discrepancy is that the translation for `Seq#Build` uses the
type of the element to decide whether to insert a box (which evaluates to
`object` in both statements), whereas the translation for `Seq#Update` uses the
`Arg` field in the type of the sequence itself to decide whether to box `a#0`.
That type evaluates to `object` in the `assume` but to `Q` in the `assert`, and
hence no box is inserted.

Closes #2429.