Fix: No crash anymore on the IDE when using the elephant operator
Fixes #2667 by remarking that sometimes, Tok is before EndTok for an obscure reason.
In the case of the elephant operator in `var x :- Parse()`, Tok designates `:-` and EndTok designates `x`, probably because the declaration is the same as the update.
I fixed the generation of faulty rangeToken like this for all statements.
Tested in the language server tests directly.