chore: Factor out projects without runtimes (#2568)
Factors out two new projects that don't depend on the `DafnyRuntime`
project or `DafnyRuntime.*` runtime artifacts, and removes the DLL-level
dependencies on `DafnyRuntime.dll`. It also has the happy side-effect of
making directory, csproj, and assembly names consistent.

The four main changes are:

* Rename `Dafny` folder/`DafnyPipeline` project to `DafnyCore`, remove
reference to `DafnyRuntime` and `DafnyRuntime.*` embedded resources
* Create new `DafnyPipeline` folder/project, referencing `DafnyCore` and
adding `DafnyRuntime.*` embedded resources
* Note `DafnyPipeline.dll` does NOT depend on `DafnyRuntime.dll`,
however! It has a project-level dependency to ensure `DafnyRuntime` is
built first, but only embeds the runtime artifacts.
* Relabel `DafnyDriver` as `DafnyDriver.dll` instead of `Dafny.dll`,
remove reference to `DafnyRuntime`
* I'll be able to use this entry point to build Dafny source code into
the runtimes when implementing #2390
* Create new `Dafny` folder/project, referencing `DafnyDriver` and
`DafnyPipeline` (and hence pulling in embedded runtimes)

Note that we are still only publishing `Dafny.dll` and
`DafnyPipeline.dll` as Nuget packages, and they still transitively
contain the same content. We could publish `DafnyDriver` and/or
`DafnyCore` in the future as well.

Removing `DafnyRuntime` as a DLL dependency from all of these projects
initially caused the `Test/dafny0/DafnyLibClient.dfy` test to break.
This is because the driver was attempting to `Assembly.Load` each DLL
passed to it, in order to read any `DafnySourceAttribute` value out of
it if present. This has never been a good idea, since an arbitrary DLL
will have dependencies not satisfied by the invoking context, and not
finding `DafnyRuntime.dll` was just one potential error this could
cause. I had previously used
[Cecil](https://www.mono-project.com/docs/tools+libraries/libraries/Mono.Cecil/)
to read DLL files without loading them, but this was lost when we moved
to .NET Core v2 since that library was not supported. This time I've
used the built-in `System.Reflection.Metadata.MetadataReader` and
related types to decode enough of the DLL to read this attribute.