Chore: Fixed HTML rendering of least and greatest predicates approximations (#2725)
This fixes a serious problem in the web documentation where the negative
space was displayed like a "!" in inline TeX and as expected in TeX
blocks. This seems to be the case because because Markdown does some
preprocessing and replaces `\!` with `!`. With the `\kern` command, we
don't have this problem.

Before:

![image](https://user-images.githubusercontent.com/3601079/189699063-94107311-5faf-4c13-8eb5-d134a37e4426.png)

After:

![image](https://user-images.githubusercontent.com/3601079/189699117-f175dec0-4b4e-40af-87e1-ad18b86de9c0.png)

20 places were concerned for "f" and 12 for "g".

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>