feat: Datatype ghost constructors (#2666)
This PR introduces ghost constructors for datatypes, as per RFC
https://github.com/dafny-lang/rfcs/pull/11.

For motivation of this feature, see the RFC or the example `Stack` class
in `Test/comp/Uninitialized.dfy`.

## Feature description

* Any (and all) constructor of a datatype can be declared as ghost. This
is done by preceding the constructor name by the keyword `ghost`. (Note,
if the constructor has attributes, those are placed before the name of
the constructor, which is consistent with the attribute placement in
other declarations. For a ghost constructor, this means the attribute is
placed between the `ghost` keyword and the identifier that names the
constructor.)
* Compiled code may not explicitly name a ghost constructor. Moreover,
compiled code is not allowed to depend on the need to check if a
datatype value is of a ghost variant. Consequently, the verifier (and in
some cases, the resolver) will enforce that any discriminator (e.g.,
`Cons?`) or destructor (e.g., `tail`) is used only when the datatype
value is known not to be of a ghost variant. In other words, a
precondition for these operations (in compiled contexts) is that the
datatype value is not of a ghost variant.
* Even though a program cannot (in compiled contexts) explicitly create
ghost variants, Dafny might assign ghost variants as part of
auto-initialization. These "assignments" are performed by... doing
nothing! (Or nearly nothing.) In other words, since the verifier makes
sure that ghost variants are never looked at, the compiler feels free to
leave such variables as uninitialized.
* Uninitialized variables usually means asking for trouble. Ghost
constructors provides uninitialized storage and verifies that that
program never looks at uninitialized variables. Still, there are two
areas where uninitialized storage may affect execution (these two points
were not mentioned in the RFC):

- If storage is really uninitialized, then a run-time garbage collector
may not be able to trace pointers accurately.
- Dafny does not restrict `print`ing of uninitialized values. If a
program uses `print` with an uninitialized values, there is in general
no way to tell what will happen.

The compilers that Dafny supports are managed languages (with the
exception of C++, but C++ does not use a garbage collector; furthermore,
the C++ compiler currently supports only a subset of Dafny, and that
subset does not include ghost constructors). Since the target languages
are managed and type safe, there is no way for the Dafny compilers to
really leave storage uninitialized. Instead, values will typically have
a `null`-equivalent value. This means that garbage collection will work
and `print`ing will not crash the program.

As for printing, `print` should be considered a debugging facility or a
facility to make it easy to write student programs.

* The RFC mentions a compiler optimization where a datatype with exactly
one non-ghost constructor with exactly one non-ghost parameter can
compiled as just that one parameter. This PR does not implement that
optimization. A separate PR could implement this optimization along with
the similar optimization for tuples with ghost components. Ghost
constructors would benefit from the optimization, but do not depend on
it.


Closes https://github.com/dafny-lang/rfcs/pull/11

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

Co-authored-by: Fabio Madge <fmadge@amazon.com>