Publish DafnyServer/Driver/Core to NuGet as well (#2787)
Fixes #2776: the root cause is that `DafnyLanguageServer` depends on
`DafnyServer`, but the latter was never published. Also publishes the
new `DafnyDriver` and `DafnyCore` packages I extracted out earlier, as
otherwise the same problem would affect more packages on the next
release.

Also replaces `version.cs` with the more automatic mechanism of a
`Build.Directory.props` file, just as Boogie uses. The side effect is
that the version is more consistently aligned across all the published
packages. In particular it changes the current `DafnyRuntime` version
from `1.2.0` to `3.8.1.40901` (or rather `3.9.0` by the time we release
this).

I managed to augment the existing GHA that verifies the `dafny` dotnet
tool to also do a very superficial verification that each library we
publish can at least resolve all of its dependencies through NuGet.