Add plugin to create wrapper functions/methods to check contracts at runtime (#2712)
Add a flag to allow certain function and method contracts to be tested
at runtime. For the moment, it focuses on `{:extern}` code, but could be
applied to other code in the future. Some details of the design are
described in the comment thread of issue #2235.

It is usable with the following:
```
/testContracts:<AllExterns|ExternsInTests>
    Enable run-time testing of compiled function or method contracts in
    certain situations, currently focused on :extern code.

    AllExterns - Check contracts on every call to a function or
        method marked :extern, regardless of where it occurs.
    Externs - Check contracts on every call to a function or method
        marked :extern when it occurs in one marked :test.
```

Fixes #2235.