feat: Rewrite the AutoExtern CLI and allow generalized name rewrites (#2849)
This PR makes AutoExtern more flexible, removing the need for a
workaround in the compiler-bootstrap project.

Description from
https://github.com/dafny-lang/compiler-bootstrap/commit/a49b299fda7479d168a8c34237a890214804e161:

The Microsoft.Dafny C# namespace is very large, and we don't model every
part of it¹. Instead, we have two models:

- One, handwritten, capturing things like `BuiltIns`, `Cloner`,
`ConcreteSyntaxTree`, `ErrorReporter`, etc.
- One, auto-generated, translating all of the AST.

This started causing issues with the introduction of `Dafny.IToken`,
because that interface is defined in the AST but used in the
`ErrorReporter`, so both of our models needed to have their own
definition of `IToken`.

Dafny doesn't support merging modules defined in different files, so
previously the two models were two different Dafny modules. They both
compiled to the same C# namespace using `:extern` annotations, but since
Dafny checks for unicity of `:extern` module names we had to use a
different name (`__AUTOGEN__`) and then run `sed` on the output.

Now we have just one module. The main trick was extending `AutoExtern`
to support renaming symbols, so that we could adjust the generated code
to refer to local symbols correctly (concretely, referring to
`Microsoft.Boogie.X` as `Boogie.X`).

¹ AutoExtern isn't good enough yet to translate all of Dafny
automatically; for example, it doesn't handle methods.

Closes #2839