Enable passing program arguments when using the new CLI format (#2823)
### Changes

Enables passing program arguments when using the `run` command of the
new CLI format. There is no `--args` option or handling of `--`
different from POSIX. Instead, any arguments after the first one, which
is interpreted as a source file, are interpreted as arguments to the
program.

This is similar to what npm and dotnet do.

We don't expect many use-cases for `run` where there is a need to pass
multiple input files, but when there is, the new option `--input <file>`
can be used. This option can be used multiple times to add multiple
additional input files.

In the future we also expect to have a project file for Dafny that'll
enable specifying all the project's source files, and then the project
file will be the single passed input file, which is similar to how
`dotnet` works.

### Testing

- Update an existing test to also test this feature

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>