feat: /runAllTests:1 MUST set the exit code on failed tests (#3047)
Fixes #3043.

Replaces just printing a "Test failures occurred: see above." message
with an `expect` statement instead, such that if there are test failures
the overall program will halt instead of terminate normally.

In addition, ensure that if a program halts when executed through a main
method (as opposed to invoked from external code) ensure the exit code
of the process is not zero. This could be considered a breaking change,
but I think it's very reasonable to call this a strict improvement in
behavior if not an outright bug fix instead, as anyone with failing
expect statements they are not aware of will almost certainly be glad to
discover them ASAP.

In combination, this ensures that when using `/runAllTests:1`, a test
failure will cause the exit code to be non-zero.

Also added support for prefixing a LIT command line with `!` to succeed
iff the exit code is non-zero. This is already a fairly portable shell
syntax and supported by the original LLVM LIT runner.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>