Add warnings about possibly confused operator precedence (#2783)
A common mistake is to write an expression like

``` dafny
&& A
&& B ==> C
&& D
```

with the intention of writing

``` dafny
&& A
&& (B ==> C)
&& D
```

This PR emits warnings when certain operator arguments seem to go on for
too long, where "seem to" is judged based on identation.

The basic idea is to look where the right-hand operand of `==>` starts
and then insist that all tokens in the RHS are aligned at, or to the
right of, the column of the RHS start. This same idea is used for other
kinds of "right-hand sides", too, like the `else` branch of an
`if-then-else` expression and the body ("term") of logical quantifiers.

When looking at the RHS, certain parts of the RHS expression are ignored
(or, rather, are treated separately). This happens when an expression
has a "middle" part, like the `then` branch of an `if-then-else`
expression. For example, no warning is produced for

``` dafny
  if x < 0 then
-x
    else x
```

If an expression of interest `e` is found in the RHS, then `e`'s
internal structure is analyzed for `e` but not for the enclosing
expression. For example, no warning is produced for

``` dafny
  forall x :: P(x) ==>
Q(x)
```

The left-hand side of `==>` is also analyzed. Some more details are
found in the `VisitOneExpr` method in `PrecedenceLinter.cs`.

The binary operators `==>`, `<==`, and `<==>` are treated as mentioned
above, since these operators have relatively low binding powers. The
linter does not do this processing for any other binary operators, like
`&&`, `||`, `+`, and `*`.

Fixes #1311

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>