Ensure non-parseable SeqIndexs do not throw on parse failure
Occasionally the language server is passed a seq index that is not parseable as
an integer.  While the root cause has not yet been determined, it's still
causing an LSP crash.  This patch ensures that non-parseable indices are handled
separately from those that are parseable.

After this patch, any non-integral indices are appended to the end of the
sorted output: the crashing test in #3048 now produces,

```
@4:seq<seq<seq<_module.uint8>>> = (Length := 9899, [0] := @14, [8100] := @15, [9898] := @16, [(- 1)] := @17)
```