Add --test-assumptions option (#3185)
### Changes

Add the `--test-assumptions` option to all execution commands:
run,build,translate,test

Note that I've written the description of the option to be aspirational,
so that it indicates what we want this option to do, with the note that
it's experimental and after that an explanation of what it currently
does. I've also simplified the option so it's just an off/on boolean,
instead of allowing using to choose under what conditions they want to
add tests, since I don't yet see a use-case for the previous
`:TestExterns` option. Feel free to discuss!

### Testing

Add the new option in a littish test

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>