Move call graph building (#3193)
This PR moves the building of the call graph out from the general name
resolution + type inference phase of the Resolver. This call-graph
building also sets the following fields (which were previously set in
the resolution/inference phase, too):

- the `.IsRecursive`, `.AllCalls`, and `.ContainsQuantifier` fields of
`Function`,
- the `.Uses` field of `ExtremePredicate`, and
- the .`IsRecursive` and `.AssignedAssumptionVariables` fields of
`Method`.

In addition, this PR fixes some minor issues, like
- some missing/confused visitor cases,
- parsing attributes after the first RHS of `:-`,
- some missing resolution and/or printing of attribtues.

Please note: It is not the purpose of this PR to make everything
beautiful in the resolver. It aspires only to move the call-graph (and
related) functionality into its own resolution phase, while not making
anything less beautiful.

This PR is a first step in preparing for an upcoming PR that refreshes
Dafny's type inference. By itself, this PR does not change type
inference; it merely moves some non-type-inference stuff out of the
current type-inference phase.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

Co-authored-by: Remy Willems <rwillems@amazon.com>