To reduce special cases, also include NonNullTypeDecl in cyclicity checks
This change makes the source code more straightforward, because it eliminates a special case. It means that NonNullTypeDecl’s are now added as lone vertices into the call-graph during resolution (whereas previously they were not added at all). As far as I can tell, this does not make any difference for the verifier; it only affects the output of /rprint.